requirements.txt
asgiref==3.6.0
certifi==2021.5.30
cloudinary==1.26.0
config==0.5.0.post0
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django
django-bootstrap3
django-cors-headers
django-heroku
django-registration
django-utils==0.0.2
django-utils-six==2.0
gunicorn==20.1.0
Pillow==9.4.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
python-decouple==3.4
pytz==2021.1
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.6
whitenoise==5.2.0
djangorestframework==3.12.4